requirements.txt
asgiref==3.3.1
astroid==2.5.1
Django==3.1.7
django-rest-framework==0.1.0
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.7.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
